Extracted MySQL concerns into MySQLManager file
